app/requirements/testing.txt
-r base.txt

pytest==5.4.3
pytest-django==3.9.0
pytest-factoryboy
factory-boy==2.12.0

